✨ feat: funding API 자동 기부, 직접 기부 구현

+++ be/funding/src/main/java/com/ieum/funding/request/FundingDonationRequestDTO.java
@@ -6,7 +6,5 @@ import lombok.Data;
 public class FundingDonationRequestDTO {
     private Long fundingId;
     private Long memberId;
-    private Long amount;
-    // 추가적인 데이터 필요
-    // ex 금액
+    private Integer amount;
 }
